+++ src/app/(auth)/signup/page.tsx
@@ -1,4 +1,4 @@
-import SignupForm from '@/components/auth/SignupForm';
+import { SignupForm } from '@/components';
 
 const SignupPage = () => {
   return <SignupForm />;
